Update CRC calculator for reserved characters.
This inverter uses the '\r' and '(' characters as indicators for command
and result termination, and result start respectively. For this reason,
if either of these characters appears in the calculated CRC, they are
simply incremented by 1.

This change adds handling for this, and removes the previous question
around that and the POP02 command which has a CRC that contains one of
these characters.