First run at entity definitions and parsing framework
All entities retuned from queries are defined in the entities module and
given a unique key that represents the entity. For each entity we also
have various fields describing it like a description, format function
etc.

We then define a dict that says which entities are returned by which
queries.

Using this entity list per query, we can then create a general parser
framework that will auto parse and format the results per entity type
for a given query.

These are some query results that are returned in a compact string
format. For these, we have additional parser functions in the entities
module.